Conflicts:
	Soheil/Soheil.Core/DataServices/Diagnostic/ActionPlanDataService.cs
	Soheil/Soheil.Core/DataServices/Diagnostic/DefectionDataService.cs
	Soheil/Soheil.Core/DataServices/PP/BlockDataService.cs
	Soheil/Soheil.Core/ViewModels/PP/Editor/PlanEditorVm.cs
	Soheil/Soheil.Dal/SoheilEdm.edmx.sql

+++ Soheil/Soheil.Core/DataServices/PP/TaskDataService.cs
@@ -112,11 +112,11 @@ namespace Soheil.Core.DataServices
 					"You can't delete this Task. It has Reports",
 					Soheil.Common.SoheilException.ExceptionLevel.Error, model.Block);
 
-			foreach (var taskReportEnt in model.TaskReports)
+			foreach (var taskReportEnt in model.TaskReports.ToArray())
 			{
 				_taskReportDataService.DeleteModel(taskReportEnt);
 			}
-			foreach (var process in model.Processes.ToList())
+			foreach (var process in model.Processes.ToArray())
 			{
 				DeleteModel(process);
 			}
@@ -139,8 +139,18 @@ namespace Soheil.Core.DataServices
 			Context.Commit();
 		}
 		//Recursive (sm & po)
-		internal void DeleteModel(Process process, bool force = false)
+		/// <summary>
+		/// Deletes a process and returns a value indicating success
+		/// <para>No commit</para>
+		/// </summary>
+		/// <param name="process"></param>
+		/// <param name="force"></param>
+		/// <returns></returns>
+		internal bool DeleteModel(Process process, bool force = false)
 		{
+			if (!force && !process.IsReportEmpty)
+				return false;
+
 			foreach (var po in process.ProcessOperators.ToArray())
 			{
 				DeleteModel(po);
@@ -149,15 +159,14 @@ namespace Soheil.Core.DataServices
 			{
 				DeleteModel(sm);
 			}
-			if (process.IsReportEmpty || force)
+			foreach (var processReport in process.ProcessReports.ToArray())
 			{
-				foreach (var processReport in process.ProcessReports.ToArray())
-				{
-					_processReportDataService.DeleteModel(processReport);
-				}
+				_processReportDataService.DeleteModel(processReport);
 			}
-			process.Task.Processes.Remove(process);
+			if (process.Task != null)
+				process.Task.Processes.Remove(process);
 			_processRepository.Delete(process);
+			return true;
 		}
 
 		internal void DeleteModel(SelectedMachine sm)
